Show the time while waiting for jobs to finish.
This is really useful for me when I wake up to a screen full of "still have jobs left in the queue", but don't have a good idea of how long that has been. Also increase the time spent between checks, in order to have fewer lines in the scrollback.