Allow configuring Batch job overrides from Codebuild
This means that setting an environment variable at Codebuild can override the environment in the Batch job definition, and hence the configuration file in the Batch jobs which run various stages.

For example, if we wanted to override the S3 bucket name used for the meta-batch stage, then we could set:

`META_BATCH__TILEQUEUE__STORE__NAME=whatever`

This allows for lightweight tweaking of the configuration for a particular run, without the heavyweight approach of forking the whole `tileops` repo to change a few variables.