Use an instance profile in compute environment config.
Previously, we've been using an IAM role for this, but it looks like EC2/spot recently got a bit stricter about this being an instance profile (which is what the documentation says it should be). This adds the extra steps to create the instance profile at the same time as the role and pass it through when creating the compute environment.